rearrange and refactor view models
