+++ data_comparator/view_models/__init__.py
@@ -1,0 +1,1 @@
+name = "view_models"
